Reverting to Eleventy v2.0.1
Eleventy v3.0.0 and the required plugins led to several breakages that I tried to find fixes for but were ultimately not worth the effort.  Ultimately, more packages are switching over to be distributed as ES modules and the available versions of several key dependencies are not all compatible yet.

This version works with both `npm run-script build` and `npm run-script serve` and does not get the same errors or warnings as the previous version, while still producing the same index.html output.